requirements.txt
fastapi
uvicorn
requests
langchain
chroma
pypdf2
